Fix 73 bledy kompilacji dla Minecraft 26.2 (Mojang API)
- MultiPlayerGameMode.attack(Player, Entity) we wszystkich modulach
- Inventory: getSelectedSlot()/setSelectedSlot(), getContainerSize(), getItem()
- ClientInput: keyPresses (Input record) zamiast forwardImpulse/leftImpulse/jumping
- ServerboundMovePlayerPacket: dodatkowy parametr horizontalCollision
- ClickGUI: mouseClicked/mouseReleased/mouseDragged/charTyped/keyPressed
  na eventach net.minecraft.client.input (MouseButtonEvent/KeyEvent/CharacterEvent)
- InputConstants: KEYSYM.getOrCreate(), isKeyDown(Window,int)
- Entity.setOnGround(), ResourceKey.identifier(), Minecraft.getFps(),
  fov() OptionInstance<Integer>, Vec3.atCenterOf, Blocks.RESPAWN_ANCHOR
- AutoTotem: offhand slot 45 (26.2)

+++ src/main/java/pl/memehub/modules/render/Zoom.java
@@ -23,7 +23,7 @@ public final class Zoom extends Module {
 		var mc = mc();
 		if (mc.options != null) {
 			oldFov = mc.options.fov().get();
-			mc.options.fov().set(fov.get());
+			mc.options.fov().set((int) fov.get());
 		}
 	}
 
@@ -31,7 +31,7 @@ public final class Zoom extends Module {
 	protected void onDisable() {
 		var mc = mc();
 		if (mc.options != null && oldFov >= 0.0) {
-			mc.options.fov().set(oldFov);
+			mc.options.fov().set((int) oldFov);
 		}
 	}
 }
